Adding Resulting models and source code
I am now adding the source code alone of both transfer learning(MobileNet) and from scratch implementation since both the models are over 100MB so it was over the GitHub limit, . As it was my first project in Deep Learning it took a lot of efforts and trial and errors. I have explored a lot of new features and functionalities while working on this project, and would continue to do so. I will make more projects similar to this to enhance my practical implementation skills.

PS: I used google drive to store my data while working on the model, as google colaboratory failed to store huge amount of data. It is advised for further users to keep it in mind while working on it.